requirements.txt
Flask==3.0.3
psycopg2-binary
djangorestframework-jwt
PyJWT==2.9.0
Flask-JWT-Extended==4.7.1